==> 01_data_exploration.ipynb <==
# Data Exploration

Was able to find data on NYT crosswords which is what we will focus on. Data has been [uploaded to Google Drive](https://drive.google.com/drive/folders/101KDJ5dl_cSPPZPBMBKIpy5-qbN6HkBN?usp=share_link), but is intentionally private (for now).

The first set of data we will look at is a collection of clues and their respective answers.

import pandas as pd

df = pd.read_csv("data/nytcrosswords.csv")
df

df[df['Word'] == 'ABRAHAMLINCOLN']

## Look for bad data
Let's look for missing data, or other weirdness we want to remove from the dataset. We can start with the obvious: empty values.


df['word_length'] = df['Word'].apply(lambda x: len(str(x).strip()))
df['clue_length'] = df['Clue'].apply(lambda x: len(str(x).strip()))

df[df['word_length'] <= 1]

df[df['clue_length'] <= 0]

print(f'Number of nulls: {len(df[df["Word"].isnull()])}')
df[df["Word"].isnull()].head()

print(f'Number of nulls: {len(df[df["Clue"].isnull()])}')
df[df["Clue"].isnull()].head()

After grepping the csv manually it turns out that these `NaN` values are actually the string "NULL" in the dataset which causes parsing issues. Something we'll have to solve with the cleanup.

So far, we know that we must:
1. Remove empty rows
2. Replace `NaN` with "NULL"

Next, let's make sure that we don't have any weird characters.

allowed_chars = r'^[A-Za-z\s\."-_!]+$'
mask = ~df['Clue'].str.contains(allowed_chars)
df[mask]

Seems like we have an encoding issue. Let's dig into that.

import chardet
with open("data/nytcrosswords.csv", 'rb') as file:
    print(chardet.detect(file.read()))

We know the encoding isn't utf-8, so let's keep trying.

with open("data/nytcrosswords.csv") as file:
    print(file)

Same problem. Found another library that might be helpful - couldn't get this installed with conda, so had to do a `pip install charamel`.

from charamel import Detector
detector = Detector()
with open("data/nytcrosswords.csv") as file:
    print(detector.probe(file.read(), top=10))

Progress! Let's try with these encodings.

test_encoding_df = pd.read_csv("data/nytcrosswords.csv")

allowed_chars = r'^[A-Za-z\s\."-_!]+$'
mask = ~test_encoding_df['Clue'].str.contains(allowed_chars)
test_encoding_df[mask]

After trying all the encodings I can think of _and_ using some offline tools I'm still stumped. Let's just manually clean up the rows we can and drop the rest.

df.at[80, 'Clue'] = 'Nickname for Jose'
df.at[396, 'Clue'] = 'What the tangent of 45° is equal to'
df.at[465, 'Clue'] = 'Hagar the Horrible\'s dog'
df.at[812, 'Clue'] = 'Animal on a Jagermeister bottle'
df.at[981, 'Clue'] = 'Bit of decor?'

# from chat gpt
df.at[1111, 'Clue'] = 'Part of the body named after Dr. Ernst Grafenberg'
df.at[1147, 'Clue'] = 'Sight on an M. C. Escher Mobius strip'
df.at[1238, 'Clue'] = 'Its anthem is "Terre de nos aieux"'
df.at[1473, 'Clue'] = 'Opera division'
df.at[1476, 'Clue'] = 'Reunion, for one'
df.at[1624, 'Clue'] = 'Magazine co-founded in 1945 by Helene Gordon'
df.at[2212, 'Clue'] = 'AZUR Cote d\'Azur'
df.at[2282, 'Clue'] = 'Eleven, en francais'
df.at[2292, 'Clue'] = ' ___ d\'Etat'
df.at[2446, 'Clue'] = 'Hagar the Horrible\'s pooch'
df.at[2557, 'Clue'] = 'Like Prince William vis-a-vis Prince Harry'
df.at[2579, 'Clue'] = 'One of a resume pair?'
df.at[2701, 'Clue'] = 'Function whose output is 45° when applied to 1'
df.at[2881, 'Clue'] = 'The singer Bjork, for one'
df.at[2937, 'Clue'] = '£ : pound sterling :: R : ___'
df.at[2940, 'Clue'] = '58 ___ Mediterraneo'
df.at[3024, 'Clue'] = 'Ile de la Cite, one end of Paris\'s Pont Neuf'
df.at[3091, 'Clue'] = 'Fish often served "meuniere"'

df.at[3254, 'Clue'] = 'Wielder of the hammer Mjölnir'
df.at[3402, 'Clue'] = 'Quinceañeras, e.g.'
df.at[3450, 'Clue'] = 'Perfect vis-à-vis good, in an aphorism'
df.at[3630, 'Clue'] = 'Risqué'
df.at[4037, 'Clue'] = 'Emily Brontë pseudonym'
df.at[4056, 'Clue'] = "Aéroport d'Orly"
df.at[4407, 'Clue'] = '"90 Day Fiancé" channel'
df.at[4459, 'Clue'] = '0, for 0°'
df.at[4618, 'Clue'] = 'Word on either side of "à"'
df.at[5558, 'Clue'] = 'Opposite of là'
df.at[5883, 'Clue'] = '365 días'
df.at[6215, 'Clue'] = 'José Olé (frozen foods brand)'
df.at[6347, 'Clue'] = 'Tête-à-tête'
df.at[6409, 'Clue'] = 'What good soufflés do'
df.at[6410, 'Clue'] = '"Liberté, égalité, fraternité" (Haiti\'s motto)'
df.at[6421, 'Clue'] = 'Les États-Unis'
df.at[6571, 'Clue'] = '"De dónde eres?" (Spanish 101 query)'

allowed_chars = r'^[A-Za-z\s\."-_!£öñàéë°íê]+$'
mask = ~df['Clue'].str.contains(allowed_chars)
print(len(df[mask]))
df[mask].head(20)

Using ChatGPT the missing characters can be identified pretty easily, but it's still a huge pain to update them and I think the special characters will make our life more difficult in the future so let's just ignore them and reconsider this in the future if we need to.

We now need to:
* Remove empty rows
* Replace NaN with "NULL"
* Remove rows with characters that we don't like

Let's do this in a new notebook and then take a fresh look at the data.

## Future cleaning to consider
* unknown characters
* "see 64-across", {See Notepad}, etc ex
