==> 03_data_exploration_take_2.ipynb <==
# Data Exploration, Take 2

We previously [explored](01_data_exploration.ipynb) and then [cleaned up](02_clean_data.ipynb) the data, let's see if there is any further processing we need to do.

import pandas as pd

df = pd.read_csv("cleaned_data/clean_1.csv", keep_default_na=False)
df

## Considerations

We saw in the previous exploration some clues like "see 64-across" or {See Notepad} - we need to figure out how we can identify these. Additionally we should search for other unexpected characters.

### Reference Clues

# match clues that reference other clues (ex. 12 Down), but not 12 Downing Street
re = r'[0-9]+[-\s]+(?:down|across)+\b'

df[df['clue'].str.contains(re)]

df = df[~df['clue'].str.contains(re)]

### See Notepad clues

df[df['clue'].str.contains(r'see notepad')]

df = df[~df['clue'].str.contains(r'see notepad')]

### Other Characters

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

So far we've identified three types of clues that we should clean up:
* Clues that use entity encoding (ex. \&epsilon;)
* Clues that have HTML tags in them (ex. <em>)
* Clues that use [_clue_]. These might be common crossword formats, but let's convert them for now.

#### Entity Encoding

entity_encoding = r'&[a-z]+;'
mask = df['clue'].str.match(entity_encoding)
df[mask]

We probably could keep these, but let's keep things simple and just ignore them for now and drop during out next cleanup pass. In the future we can come back and consider handling these.

#### HTML Tags

html_tags = r'<[a-z]+>'
mask = df['clue'].str.match(html_tags)
df[mask]

Same story here - we could either keep or convert these, but let's just ignore them for now and drop during our next cleanup pass.

#### [Clues]

bracket_clues = r'\[.+\]'
mask = df['clue'].str.match(bracket_clues)
df[mask]

Again, a small number of clues so let's just ignore these for now.

#### Quick clenaup
Let's drop all of these rows and see where we're at.

entity_encoding = r'&[a-z]+;'
df = df[~df['clue'].str.contains(entity_encoding)]

html_tags = r'<[a-z]+>'
df = df[~df['clue'].str.contains(html_tags)]

bracket_clues = r'\[.+\]'
df = df[~df['clue'].str.contains(bracket_clues)]

#### Second pass

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### Starred Clues
"Starred clues" (ex. *stay in power) are common in crosswords. We should just drop the asterisk (but keep clues that are only an asterisk).

df['clue'] = df['clue'].apply(lambda x: x.lstrip('*') if x != '*' else x)

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### @ Clues

at_clues = r'@'
mask = df['clue'].str.match(at_clues)
df[mask]

These look fine, let's just allow them.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### () Clues

bracket_clues = r'\(|\)'
mask = df['clue'].str.match(bracket_clues)
df[mask]

These look fine, let's just allow them.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### # Clues

pound_clues = r'#'
mask = df['clue'].str.match(pound_clues)
df[mask]

These look fine, let's just allow them.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### = Clues

equals_clues = r'='
mask = df['clue'].str.match(equals_clues)
df[mask]

These look good, let's allow them.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### % Clues

percent_clues = r'%'
mask = df['clue'].str.match(percent_clues)
df[mask]

Also, fine.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### Entity Encoding Take 2

We missed clues that have characters like \&#119070; &#119070;

entity_encoding = r'&#[0-9]+;'
mask = df['clue'].str.match(entity_encoding)
df[mask]

Let's drop these

entity_encoding = r'&#[0-9]+;'
df = df[~df['clue'].str.contains(entity_encoding)]

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

slash_clues = r'/|\\'
mask = df['clue'].str.match(slash_clues)
pd.set_option('display.max_rows', 11)
df[mask]

These are all good.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$/]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### Leading +

leading_plus = r'^\+'
mask = df['clue'].str.match(leading_plus)
df[mask]

Some of these seem valid, some of these we want to strip.

leading_plus = r'^\+[a-z]'
mask = df['clue'].str.match(leading_plus)
pd.set_option('display.max_rows', 15)
df[mask]

Let's strip the ones that have a leading plus, similar to what we did for the asterisk

import re

df['clue'] = df['clue'].apply(lambda x: x.lstrip('+') if re.match(leading_plus, x) else x)

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$/\+]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

##### < clues >

leading_angle = r'^<\s.+'
mask = df['clue'].str.match(leading_angle)
df[mask]

These are referencing the clue number which we don't have, so let's drop these. Interesting one to consider in the future though!

leading_angle = r'^<\s.+'
df = df[~df['clue'].str.contains(leading_angle)]

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$/\+]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

The rest of these are all pretty nuanced, let's drop them for now and we can always revisit later.

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$/\+]+'
df = df[df['clue'].str.contains(allowed_chars)]

allowed_chars = r'^[a-z0-9\s\."\_!\-\'\@\(\)#=%$/\+]+'
mask = ~df['clue'].str.match(allowed_chars)
df[mask]

df

We now have what we think is a clean dataset! First, let's create a new notebook to make this cleanup more clear and then we can create a new CSV with the cleaned data

### Answers
Let's make sure that all the answers are only letters.

df[~df['answer'].str.contains(r'^[a-z]+$')]

These don't make sense in a crossword, let's remove them

df = df[df['answer'].str.contains(r'^[a-z]+$')]

df

## Future Considerations
* Encode special characters in a standard way
* Cosnider nuances of crosswords - < reference clues, cross reference clues, asterisk clues, etc.
