==> 05_first_neural_net.ipynb <==
# First Neural Net
Let's figure out how we would take our data and use it to train a neural net. We'll focus on figuring out how to tokenize the data and how it would be fed into and out of the network

import numpy as np
import pandas as pd
import re
import torch
import torch.nn.functional as F

df = pd.read_csv("cleaned_data/clean_2.csv", keep_default_na=False)

## Vocabulary Building
Let's use pytorch to build up a vocabulary. The default tokenizer is probably fine for now, but something we should consider changing in the future.

from torchtext.data.utils import get_tokenizer
from torchtext.vocab import build_vocab_from_iterator

tokenizer = get_tokenizer('basic_english')

def yield_tokens(data_iter):
    for text in data_iter:
        yield tokenizer(text)

answers_vocab = build_vocab_from_iterator(yield_tokens(df['answer'].values), specials=["<unk>"])
clues_vocab = build_vocab_from_iterator(yield_tokens(df['clue'].values), specials=["<unk>"])

print('Tokens in clues vocab:', len(clues_vocab))
print('Tokens in answers vocab:', len(answers_vocab))

Add some helpers to make conversion from strings to tokens easy:

clue_pipeline = lambda x: clues_vocab(tokenizer(x))

clue_pipeline('capital of canada')

answers_vocab.get_stoi()['ottawa']

## Data Setup
Let's create tensors for our inputs and labels.

### Input and Output
We want to input the clue and output the potential answer. Clues are encoded as a list of ints where each int ia a word in the clue vocabulary; answers are a single int which represents a word in the answer vocabulary.

We can directly input the encoded clue, but we need to consider that clues are different lengths and we need a consistent shape for our inputs. We can solve for this by padding inputs to a fixed length.

First, we must add a padding token to our vocabulary. Second, we need to figure out the length of the input which we can base off of the longest clue in the training set (note: this will fail if any clues in the test set are longer).

answers_vocab = build_vocab_from_iterator(yield_tokens(df['answer'].values), specials=["<pad>"])
clues_vocab = build_vocab_from_iterator(yield_tokens(df['clue'].values), specials=["<pad>"])

print(answers_vocab.get_itos()[0], clues_vocab.get_itos()[0])

PADDING_TOKEN_INDEX = 0

### Data Splits
Let's create train, dev, and test splits of the data. We'll use a 80-10-10 breakdown

train, dev, test = np.split(df.sample(frac=1, random_state=42), [int(.8*len(df)), int(.9*len(df))])

# Print the size of each subset
print(f'Train set size: {len(train)}')
print(f'Dev set size: {len(dev)}')
print(f'Test set size: {len(test)}')

train.sort_index()

Let's figure out what the longest clue is

train['clue'].apply(lambda x: len(clue_pipeline(x))).max()

dev['clue'].apply(lambda x: len(clue_pipeline(x))).max()

test['clue'].apply(lambda x: len(clue_pipeline(x))).max()

Let's add some buffer to reduce chances that we will see a longer clue in the future.

PAD_TO_SIZE = 45

### Tensor Building
Let's build X and Y tensors for our splits

def build_dataset(split):
    splits = {
        'train': train,
        'dev': dev,
        'test': test
    }
    df = splits[split]
    answers_stoi = answers_vocab.get_stoi()
    
    X = []
    Y = []

    for clue in df['clue'].values:
        indicies = clue_pipeline(clue)
        indicies += [PADDING_TOKEN_INDEX] * (PAD_TO_SIZE - len(indicies))
        X.append(indicies)

    for answer in df['answer'].values:
        answer_index = answers_stoi[answer]
        Y.append(answer_index)
        
    X = torch.tensor(X)
    Y = torch.tensor(Y)
    return (X, Y)


Xdev, Ydev = build_dataset('dev')
Xtest, Ytest = build_dataset('test')
Xtr, Ytr = build_dataset('train')

print(Xtr.shape, Ytr.shape)
print(Xdev.shape, Ydev.shape)
print(Xtest.shape, Ytest.shape)

## Neural Network
Let's build a simple network based on the MLP discussed in [this lecture](https://github.com/karpathy/nn-zero-to-hero/blob/master/lectures/makemore/makemore_part2_mlp.ipynb).

We will feed the clues into an embedding layer which will provide the inputs to a set of neurons which will output the liklihood of each item in the answers vocabulary being correct.

# hyperparameters
EMBEDDING_DIMENSIONS = 25
HIDDEN_NEURON_COUNT = 200
BATCH_SIZE = 64

g = torch.Generator().manual_seed(42)
C = torch.randn((len(clues_vocab), EMBEDDING_DIMENSIONS))
W1 = torch.randn(PAD_TO_SIZE * EMBEDDING_DIMENSIONS, HIDDEN_NEURON_COUNT)
b1 = torch.randn(HIDDEN_NEURON_COUNT)
W2 = torch.randn((HIDDEN_NEURON_COUNT, len(answers_vocab)))
b2 = torch.randn(len(answers_vocab))
parameters = [C, W1, b1, W2, b2]
for p in parameters:
    p.requires_grad = True

sum(p.nelement() for p in parameters)

Let's construct the forward pass and make sure we understand what's happening.

for i in range(1):
    # minibatch
    ix = torch.randint(0, Xtr.shape[0], (BATCH_SIZE,))
    
    # forward pass
    
    # do lookup
    emb = C[Xtr[ix]] # (BATCH_SIZE, PAD_TO_SIZE, EMBEDDING_DIMENSIONS)
    print(emb.shape)
    print(f'{BATCH_SIZE},{PAD_TO_SIZE}, {EMBEDDING_DIMENSIONS}')
    
    # concat embeddings together
    concat = emb.view(-1, PAD_TO_SIZE * EMBEDDING_DIMENSIONS) # (BATCH_SIZE, PAD_TO_SIZE * EMBEDDING_DIMENSIONS)
    print(concat.shape)
    print(f'{BATCH_SIZE},{PAD_TO_SIZE * EMBEDDING_DIMENSIONS}')
    
    # hidden layer
    h = torch.tanh(concat @ W1 + b1) # (BATCH_SIZE, HIDDEN_NEURON_COUNT)
    print(h.shape)
    print(f'{BATCH_SIZE},{HIDDEN_NEURON_COUNT}')
    
    # output layer
    logits = h @ W2 + b2 # (BATCH_SIZE, len(answers_vocab))
    print(logits.shape)
    print(f'{BATCH_SIZE},{len(answers_vocab)}')
    
    loss = F.cross_entropy(logits, Ytr[ix])
    print(loss.item())

Looks good! Loss is high, but we'll fix that 😅. Let's start with a simple training cycle to see if what we have works.

%%time
LEARNING_RATE = 0.1

for i in range(1000):
    # minibatch
    ix = torch.randint(0, Xtr.shape[0], (BATCH_SIZE,))
    
    # forward pass
    emb = C[Xtr[ix]] # (BATCH_SIZE, PAD_TO_SIZE, EMBEDDING_DIMENSIONS)
    concat = emb.view(-1, PAD_TO_SIZE * EMBEDDING_DIMENSIONS) # (BATCH_SIZE, PAD_TO_SIZE * EMBEDDING_DIMENSIONS)
    h = torch.tanh(concat @ W1 + b1) # (BATCH_SIZE, HIDDEN_NEURON_COUNT)
    logits = h @ W2 + b2 # (BATCH_SIZE, len(answers_vocab))
    loss = F.cross_entropy(logits, Ytr[ix])
    if i == 0:
        print('Initial loss:', loss.item())
    
    # backward pass
    for p in parameters:
        p.grad = None
    loss.backward()
    
    # update
    for p in parameters:
        p.data += -LEARNING_RATE * p.grad

print('Final loss:', loss.item())

Seems, to be working! Now let's figure out how to optimize our hyperparameters. The first step is to evaluate our result against the dev set.

%%time
emb = C[Xdev] # (BATCH_SIZE, PAD_TO_SIZE, EMBEDDING_DIMENSIONS)
concat = emb.view(-1, PAD_TO_SIZE * EMBEDDING_DIMENSIONS) # (BATCH_SIZE, PAD_TO_SIZE * EMBEDDING_DIMENSIONS)
h = torch.tanh(concat @ W1 + b1) # (BATCH_SIZE, HIDDEN_NEURON_COUNT)
logits = h @ W2 + b2 # (BATCH_SIZE, len(answers_vocab))
loss = F.cross_entropy(logits, Ydev)

print('Dev loss:', loss.item())

We have a solid setup for tuning our hyperparameters - let's learn how to do that in the next notebook.
