==> 11_duplicate_data.ipynb <==
# Duplicate Data
It turns out that we have quite a bit of duplicate data in our dataset which is making training less useful and likely over estimating our model's effectiveness because of repeated data between the test and train datasets.

Let's remove the duplicates.

import pandas as pd
df = pd.read_csv("cleaned_data/clean_2.csv", keep_default_na=False)
df

dropped_duplicates = df.drop_duplicates()
print(f'Dropped {len(df) - len(dropped_duplicates)} duplicates')
print(f'{len(df)=},{len(dropped_duplicates)=}')

dropped_duplicates.to_csv('cleaned_data/clean_3.csv', index=False)
