==> 12_long_input.ipynb <==
# Long Input
At some point when training I realized taht we were bloating the network size because of a few clues containing a long number of words. We were padding every input to length 40 and each of those were going into the embedding table. Let's take a look at how many long clues we actually have. 

import pandas as pd
df = pd.read_csv("cleaned_data/clean_2.csv", keep_default_na=False)
df

from torchtext.data.utils import get_tokenizer
tokenizer = get_tokenizer('basic_english')

df['clue_tokens'] = df['clue'].apply(lambda x: len(tokenizer(x)))
df.sort_values(by='clue_tokens', ascending=False)

Interesting that some clues have zero tokens...

df[df['clue_tokens'] == 0]

len(df) - len(df[df['clue_tokens'] <= 10])

So, if we remove all the clues longer than 10 tokens we can shrink our network without losing much data.

print(len(df))
df = df[df['clue_tokens'] > 0]
print(len(df))
df = df[df['clue_tokens'] <= 10]
print(len(df))

df = df.drop(columns=['clue_tokens'])
df.to_csv('cleaned_data/dupes_10_or_less_tokens.csv', index=False)
