==> 16_char_level_model.ipynb <==
# Character Level Model
We didn't see much success with our previous models. We topped out at about 40% accuracy even having duplicates in the dataset.

My theory is that the model struggles because the input and output spaces are so large: ~77k input tokens ~60k output tokens. This means the network must be quite large which makes it difficult to train.

Next, I'd like to recondiser this problem as a character level problem. Given the answer,clue pair `neural,kind of network` we can consider the problem on a character by character basis:

answer = 'neural'
clue = 'kind of network'

print('||||| ', end='')

for c in clue:
    print(f'{c} --> ', end='')

print('||||| ', end='')

for c in answer:
    print(f'{c} --> ', end='')

print('||||| ')

Now, let's replace those `|||||` separators with things that are meaningful. We'll use `<BOC>`, `<EOC>`, and `<EOA>` tokens.

tokens = ['<BOC>']
tokens.extend(list(answer))
tokens.append('<EOC>')
tokens.extend(list(answer))
tokens.append('<EOA>')

for i, token in enumerate(tokens):
    print(f'{token}', end='')
    if i < len(tokens) - 1:
        print(' --> ', end='')

So that looks good, but let's think about the `<EOC>` token - what if we replaced that with a special token indicating the _length of the expected answer_. This serves as a useful indicator to the network as to what should be generated next.

tokens = ['<BOC>']
tokens.extend(list(clue))
tokens.append('<' + str(len(answer)) + '>')
tokens.extend(list(answer))
tokens.append('<EOA>')

for i, token in enumerate(tokens):
    print(f'{token}', end='')
    if i < len(tokens) - 1:
        print(' --> ', end='')

## Working with our Dataset
Let's load up a few data rows and see how this would work

import pandas as pd

df = pd.read_csv("cleaned_data/dupes_10_or_less_tokens.csv", keep_default_na=False)
df.head(5)

def build_tokens(row):
    clue = row['clue']
    answer = row['answer']
    tokens = ['<BOC>']
    tokens.extend(list(clue))
    tokens.append('<' + str(len(answer)) + '>')
    tokens.extend(list(answer))
    tokens.append('<EOA>')
    return tokens

token_batch = []
for i in range(5):
    tokens = build_tokens(df.iloc[i])
    token_batch.append(tokens)
    print(tokens)

That looks good, let's try this out with the first few rows in our dataset.

rows = df.iloc[:100]
all_chars = ''.join(rows['answer']) + ''.join(rows['clue'])

vocab = sorted(list(set(all_chars)))
vocab += ['<BOC>', '<EOA>']

max_answer_length = rows['answer'].str.len().max()
for i in range(1, max_answer_length + 1):
    vocab.append(f'<{i}>')

vocab_stoi = {s:i+1 for i,s in enumerate(vocab)}
vocab_itos = {i:s for s,i in vocab_stoi.items()}
vocab_size = len(vocab)
vocab_size

Now that we have our vocabulary built let's build a model for it.

import torch
import torch.nn as nn

class TextGenerator(nn.Module):
    def __init__(self, vocab_size, hidden_size, n_layers=1):
        super(TextGenerator, self).__init__()

        # identiy matrix for generating one-hot vectors
        self.ident = torch.eye(vocab_size)

        # recurrent neural network
        self.rnn = nn.GRU(vocab_size, hidden_size, n_layers, batch_first=True)

        # a fully-connect layer that outputs a distribution over
        # the next token, given the RNN output
        self.decoder = nn.Linear(hidden_size, vocab_size)
    
    def forward(self, inp, hidden=None):
        inp = self.ident[inp]                  # generate one-hot vectors of input
        output, hidden = self.rnn(inp, hidden) # get the next output and hidden state
        output = self.decoder(output)          # predict distribution over next tokens
        return output, hidden

model = TextGenerator(vocab_size, 64)

Let's see what our network will produce given the start token.

criterion = nn.CrossEntropyLoss()
bos_input = torch.Tensor([vocab_stoi['<BOC>']]).long().unsqueeze(0)
output, hidden = model(bos_input, hidden=None)
output # distribution over the first token

first_clue = rows['clue'][0]
first_answer = rows['answer'][0]
target = torch.Tensor([vocab_stoi[first_clue[0]]]).long().unsqueeze(0)
criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
          target.reshape(-1))             # reshape to 1D tensor

Instead of passing this token back to the NN to genereate the next one, we actually pass in the ground truth data and see the results.

# Use teacher-forcing: we pass in the ground truth `target`,
# rather than using the NN predicted distribution
output, hidden = model(target, hidden)
output # distribution over the second token

target = torch.Tensor([vocab_stoi[first_clue[1]]]).long().unsqueeze(0)
criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
          target.reshape(-1))             # reshape to 1D tensor

for i in range(2, len(first_clue)):
    output, hidden = model(target, hidden)
    target = torch.Tensor([vocab_stoi[first_clue[i]]]).long().unsqueeze(0)
    loss = criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
                     target.reshape(-1))             # reshape to 1D tensor
    print(i, output, loss)

Now, we'd expected the answer length token.

answer_token = f'<{len(first_answer)}>'
answer_token

output, hidden = model(target, hidden)
target = torch.Tensor([vocab_stoi[answer_token]]).long().unsqueeze(0)
loss = criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
                 target.reshape(-1))             # reshape to 1D tensor
print(i, output, loss)

Then, on to the answer

for i in range(len(first_answer)):
    output, hidden = model(target, hidden)
    target = torch.Tensor([vocab_stoi[first_answer[i]]]).long().unsqueeze(0)
    loss = criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
                     target.reshape(-1))             # reshape to 1D tensor
    print(i, output, loss)

Finally, we'd expect the `<EOA>` token.

output, hidden = model(target, hidden)
target = torch.Tensor([vocab_stoi["<EOA>"]]).long().unsqueeze(0)
loss = criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
                 target.reshape(-1))             # reshape to 1D tensor
print(i, output, loss)

We don't actually need to loop like this though, we can just pass in the sequence:

indices = [vocab_stoi[ch] for ch in token_batch[0]]
tensor = torch.Tensor(indices).long().unsqueeze(0)

print(tensor.shape)

output, hidden = model(tensor[:,:-1]) # <EOS> is never an input token
target = tensor[:,1:]                 # <BOS> is never a target token
loss = criterion(output.reshape(-1, vocab_size), # reshape to 2D tensor
                 target.reshape(-1))             # reshape to 1D tensor

optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
criterion = nn.CrossEntropyLoss()
for it in range(500):
    optimizer.zero_grad()
    output, _ = model(tensor[:,:-1])
    loss = criterion(output.reshape(-1, vocab_size),
                 target.reshape(-1))
    loss.backward()
    optimizer.step()

    if (it+1) % 100 == 0:
        print("[Iter %d] Loss %f" % (it+1, float(loss)))

## Full dataset

def build_tokens(row):
    clue = row['clue']
    answer = row['answer']
    tokens = ['<BOC>']
    tokens.extend(list(clue))
    tokens.append('<EOC>')
    tokens.append('<' + str(len(answer)) + '>')
    tokens.extend(list(answer))
    tokens.append('<EOA>')
    return tokens

token_batch = []
for i in range(5000):
    tokens = build_tokens(df.iloc[i])
    token_batch.append(tokens)

rows = df.iloc[:5000]
all_chars = ''.join(rows['answer']) + ''.join(rows['clue'])

vocab = sorted(list(set(all_chars)))
vocab += ['<BOC>', '<EOC>', '<EOA>']

max_answer_length = rows['answer'].str.len().max()
for i in range(1, max_answer_length + 1):
    vocab.append(f'<{i}>')

vocab_stoi = {s:i for i,s in enumerate(vocab)}
vocab_itos = {i:s for s,i in vocab_stoi.items()}
vocab_size = len(vocab)
vocab_size


def train(model, data, batch_size=1, num_epochs=1, lr=0.001, print_every=100):
    optimizer = torch.optim.Adam(model.parameters(), lr=lr)
    criterion = nn.CrossEntropyLoss()
    it = 0
    
    for e in range(num_epochs):
        # get training set
        avg_loss = 0

        for token in data:
            indicies = torch.tensor(list(map(lambda x: vocab_stoi[x], token)))
            target = indicies[1:]
            inp = indicies[:-1]
            # target = token[:, 1:] # skip the first character as that's not a target
            # inp = token[:, :-1] # skil the last value as it's never an input
            # cleanup
            optimizer.zero_grad()
            # forward pass
            output, _ = model(inp)
            loss = criterion(output.reshape(-1, vocab_size), target.reshape(-1))
            # backward pass
            loss.backward()
            optimizer.step()

            avg_loss += loss
            it += 1 # increment iteration count
            if it % print_every == 0:
                print("[Iter %d] Loss %f" % (it+1, float(avg_loss/print_every)))
                print(token)
                #print(indicies.shape)
                #print(output.shape)
                predicted_indicies = output.argmax(1)
                #print(predicted_indicies.shape)
                print(list(map(lambda x: vocab_itos[x.item()], predicted_indicies)))
                # get answer length, would be nice to have this beforehand
                #print("    " + sample_sequence(model, 140, 0.8))
                avg_loss = 0

model = TextGenerator(vocab_size, 128)
train(model, token_batch, batch_size=1, num_epochs=1, lr=0.003, print_every=1000)

## Open Questions
* Should we use separate vocabs for answers??

## References
* https://www.cs.toronto.edu/~lczhang/360/lec/w08/gen.html
